pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["metrics", "models", "networks", "pipelines", "utils"]

[project]
name = "diffusion-ts"
version = "0.1.0"
requires-python = ">=3.10"
description = "Time series generation with diffusion models"

dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "ruamel-yaml>=0.18.0",
    "wandb>=0.25.0",
    "optuna>=4.7.0",
    "optuna-integration>=4.7.0",
    "einops>=0.8.2",
    "torchmetrics>=1.9.0",
    "lightning>=2.6.1",
    "diffusers>=0.37.0",
    "numpy>=2.2.6",
    "scipy>=1.15.3",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "plotly>=6.6.0",
    "ipython>=8.38.0",
    "tqdm>=4.67.3",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "torchaudio>=2.10.0",
]
cu118 = [
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "torchaudio>=2.7.1",
]
cu124 = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "torchaudio>=2.6.0",
]
dev = [
    "ruff>=0.15.5",
    "pytest>=9.0.2",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
]

[tool.uv]
conflicts = [
    [{ extra = "cpu" }, { extra = "cu118" }, { extra = "cu124" }],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu",  extra = "cpu"   },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu124", extra = "cu124" },
]
torchvision = [
    { index = "pytorch-cpu",  extra = "cpu"   },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu124", extra = "cu124" },
]
torchaudio = [
    { index = "pytorch-cpu",  extra = "cpu"   },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu124", extra = "cu124" },
]

[[tool.uv.index]]
name     = "pytorch-cpu"
url      = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name     = "pytorch-cu118"
url      = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name     = "pytorch-cu124"
url      = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.ruff]
line-length = 100
